requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httptools==0.6.4
idna==3.10
multidict==6.4.4
packaging==25.0
peewee==3.18.1
pefile==2023.2.7
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.1
pydantic==2.11.5
pydantic_core==2.33.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
python-dotenv==1.1.0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
